Move executor tests to their own file
